wip: inital back-end implementaiton with no tests

+++ packages/api-apw/src/storageOperations/changeRequestStorageOperations.ts
@@ -88,7 +88,7 @@ export const createChangeRequestStorageOperations = (
             const existingEntry = await getChangeRequest({ id: params.id });
 
             /**
-             * Only owner can update the change request
+             * Only creator can update the change request
              */
             if (existingEntry.createdBy.id !== security?.getIdentity()?.id) {
                 throw new WebinyError(
@@ -111,8 +111,22 @@ export const createChangeRequestStorageOperations = (
         },
         async deleteChangeRequest(params) {
             const model = await getChangeRequestModel();
-
             if (security.getIdentity()) {
+                /**
+                 * We're fetching the existing entry here because we're not accepting "app" field as input,
+                 * but, we still need to retain its value after the "update" operation.
+                 */
+                const existingEntry = await getChangeRequest({ id: params.id });
+
+                /**
+                 * Only creator can update the change request
+                 */
+                if (existingEntry.createdBy.id !== security?.getIdentity()?.id) {
+                    throw new WebinyError(
+                        "Could not update the change request. Only owner can update the change request"
+                    );
+                }
+
                 await security.withoutAuthorization(async () => {
                     return cms.deleteEntry(model, params.id);
                 });
